Fix: move view page to ui folder
